confuse_ios/requirements.txt

# chardet==3.0.4
# charset-normalizer==3.1.0
# DingtalkChatbot==1.5.7
# distlib==0.3.6
# filelock==3.12.0
# gitdb==4.0.10
# GitPython==3.1.31
# idna==2.10
# smmap==5.0.0
# urllib3==1.25.11
# certifi
# requests==2.24.0
# platformdirs==3.2.0
# pbxproj               3.5.0


certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
idna==2.10
pip==21.2.4
platformdirs==3.2.0
requests==2.24.0
setuptools==58.0.4
tinify==1.6.0
urllib3==1.25.11
virtualenv==20.23.0
Pillow==10.0.0
pbxproj==3.5.0
lxml==4.9.3
guizero==1.5.0
pyinstaller==6.4.0

